requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.8.0
scipy>=1.7.0
networkx>=2.6.0
plotly>=5.0.0
wandb>=0.12.0
jupyter>=1.0.0
ipykernel>=6.0.0
pyyaml>=6.0
argparse
logging
pathlib
typing
dataclasses
